+++ example/lib/page/developer/develop_index_page.dart
@@ -267,7 +267,7 @@ class _DeveloperIndexPageState extends State<DeveloperIndexPage> {
       final assets = await PhotoManager.editor.darwin.saveLivePhoto(
         imageFile: imgFile,
         videoFile: videoFile,
-        filename: 'preview_0',
+        title: 'preview_0',
       );
       print('save live photo result : ${assets?.id}');
     } finally {