+++ ios/Classes/core/PMManager.h
@@ -21,7 +21,7 @@ typedef void (^ChangeIds)(NSArray<NSString *> *);
 #define PM_IMAGE_CACHE_PATH @".image"
 #define PM_FULL_IMAGE_CACHE_PATH @"flutter-images"
 
-typedef void (^AssetResult)(PMAssetEntity *);
+typedef void (^AssetBlockResult)(PMAssetEntity *, NSObject *);
 
 
 @interface PMManager : NSObject
@@ -44,9 +44,9 @@ typedef void (^AssetResult)(PMAssetEntity *);
 
 - (void)clearCache;
 
-- (void)getThumbWithId:(NSString *)id option:(PMThumbLoadOption *)option resultHandler:(NSObject <PMResultHandler> *)handler progressHandler:(NSObject <PMProgressHandlerProtocol> *)progressHandler;
+- (void)getThumbWithId:(NSString *)assetId option:(PMThumbLoadOption *)option resultHandler:(NSObject <PMResultHandler> *)handler progressHandler:(NSObject <PMProgressHandlerProtocol> *)progressHandler;
 
-- (void)getFullSizeFileWithId:(NSString *)id isOrigin:(BOOL)isOrigin subtype:(int)subtype resultHandler:(NSObject <PMResultHandler> *)handler progressHandler:(NSObject <PMProgressHandlerProtocol> *)progressHandler;
+- (void)getFullSizeFileWithId:(NSString *)assetId isOrigin:(BOOL)isOrigin subtype:(int)subtype resultHandler:(NSObject <PMResultHandler> *)handler progressHandler:(NSObject <PMProgressHandlerProtocol> *)progressHandler;
 
 - (PMAssetPathEntity *)fetchPathProperties:(NSString *)id type:(int)type filterOption:(NSObject<PMBaseFilter> *)filterOption;
 
@@ -55,45 +55,45 @@ typedef void (^AssetResult)(PMAssetEntity *);
 - (void)saveImage:(NSData *)data
             filename:(NSString *)filename
             desc:(NSString *)desc
-            block:(AssetResult)block;
+            block:(AssetBlockResult)block;
 
 - (void)saveImageWithPath:(NSString *)path
             filename:(NSString *)filename
             desc:(NSString *)desc
-            block:(void (^)(PMAssetEntity *))block;
+            block:(AssetBlockResult)block;
 
 - (void)saveVideo:(NSString *)path
             filename:(NSString *)filename
             desc:(NSString *)desc
-            block:(AssetResult)block;
+            block:(AssetBlockResult)block;
 
 - (void)saveLivePhoto:(NSString *)imagePath
             videoPath:(NSString *)videoPath
-            filename:(NSString *)filename
+            title:(NSString *)title
             desc:(NSString *)desc
-            block:(AssetResult)block;
+            block:(AssetBlockResult)block;
 
 - (BOOL)existsWithId:(NSString *)assetId;
 
-- (BOOL)entityIsLocallyAvailable:(NSString *)assetId resource:(PHAssetResource *)resource isOrigin:(BOOL)isOrigin;
+- (BOOL)entityIsLocallyAvailable:(NSString *)assetId resource:(PHAssetResource *)resource isOrigin:(BOOL)isOrigin subtype:(int)subtype;
 
 - (NSString*)getTitleAsyncWithAssetId:(NSString *)assetId subtype:(int)subtype;
 
 - (NSString*)getMimeTypeAsyncWithAssetId: (NSString *) assetId;
 
-- (void)getMediaUrl:(NSString *)assetId resultHandler:(NSObject <PMResultHandler> *)handler;
+- (void)getMediaUrl:(NSString *)assetId resultHandler:(NSObject <PMResultHandler> *)handler progressHandler:(NSObject <PMProgressHandlerProtocol> *)progressHandler;
 
 - (NSArray<PMAssetPathEntity *> *)getSubPathWithId:(NSString *)id type:(int)type albumType:(int)albumType option:(NSObject<PMBaseFilter> *)option;
 
-- (void)copyAssetWithId:(NSString *)id toGallery:(NSString *)gallery block:(void (^)(PMAssetEntity *entity, NSString *msg))block;
+- (void)copyAssetWithId:(NSString *)id toGallery:(NSString *)gallery block:(void (^)(PMAssetEntity *entity, NSObject *msg))block;
 
-- (void)createFolderWithName:(NSString *)name parentId:(NSString *)id block:(void (^)(NSString *, NSString *))block;
+- (void)createFolderWithName:(NSString *)name parentId:(NSString *)id block:(void (^)(NSString *newId, NSObject *error))block;
 
-- (void)createAlbumWithName:(NSString *)name parentId:(NSString *)id block:(void (^)(NSString *, NSString *))block;
+- (void)createAlbumWithName:(NSString *)name parentId:(NSString *)id block:(void (^)(NSString *newId, NSObject *error))block;
 
-- (void)removeInAlbumWithAssetId:(NSArray *)id albumId:(NSString *)albumId block:(void (^)(NSString *))block;
+- (void)removeInAlbumWithAssetId:(NSArray *)id albumId:(NSString *)albumId block:(void (^)(NSObject *error))block;
 
-- (void)removeCollectionWithId:(NSString *)id type:(int)type block:(void (^)(NSString *))block;
+- (void)removeCollectionWithId:(NSString *)id type:(int)type block:(void (^)(NSObject *))block;
 
 - (BOOL)favoriteWithId:(NSString *)id favorite:(BOOL)favorite;